% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brown_tree_prior1.R
\name{brown_tree_prior2}
\alias{brown_tree_prior2}
\title{Prior probability for the number of crossing on phylogenetic tree}
\usage{
brown_tree_prior2(phy, thed = 0, epsilon, root.prior, up = T,
  nsize = 1000, k = 0)
}
\arguments{
\item{phy}{A phylo type object.}

\item{thed}{the threshold value for Brownian motion threshold model.}

\item{epsilon}{the radius of the interval for crossing.}

\item{root.prior}{A numeric vector with two values, which is the mean and variance for the prior distribution (normal) for the root liability.}

\item{up}{Previous crossing direction. Set as TRUE for defaulted value.}

\item{nsize}{number of simulations for Monte-Carlo integral}
}
\value{
A vector associate with the prior probabililty for the number of upcrossing
}
\description{
Prior probability for the number of crossing on phylogenetic tree
}

