% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brown_tree_prior1.R
\name{brown_tree_prior_node}
\alias{brown_tree_prior_node}
\title{Generating libility value on each node of the phylogenetic tree}
\usage{
brown_tree_prior_node(phy, root.prior)
}
\arguments{
\item{phy}{A phylo type object.}

\item{up}{Previous crossing direction. Set as TRUE for defaulted value.}

\item{nsize}{number of simulations for Monte-Carlo integral}
}
\value{
A vector associate with the prior probabililty for the number of upcrossing
}
\description{
Generating libility value on each node of the phylogenetic tree
}

