% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMI_thres_posterior.R
\name{brownian_crossing_post2}
\alias{brownian_crossing_post2}
\title{Posterior probability for the number of crossing given tip binary observation}
\usage{
brownian_crossing_post2(phy, d, thed, epsilon, root.prior, ngen = 20000,
  burnin = 10000, thin = 5, k = 0, less = F)
}
\arguments{
\item{phy}{A phylo type object.}

\item{d}{the discrete observation on the tip of the tree.}

\item{epsilon}{the radius of the interval for crossing}

\item{root.prior}{A numeric vector with two values, which is the mean and variance for the prior distribution (normal) for the root liability.}
}
\value{
A vector associate with the posterior probabililty for the number of upcrossing
}
\description{
Posterior probability for the number of crossing given tip binary observation
}

