% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossing.R
\name{crossing}
\alias{crossing}
\title{Counting the number of Crossing of a process}
\usage{
crossing(pth, thed = 0, epsilon = 0, prev = 0, t = 1)
}
\arguments{
\item{thed}{the threshold value}

\item{epsilon}{half length of the interval of the threshold, that is (thed-epsilon,thed+epsilon)}

\item{path}{a vector of a stochastic process over a discrete grid}
}
\value{
crossing: the number of crossing for a interval of the threshold
}
\description{
Counting the number of Crossing of a process
}
\author{
Mingwei Tang
}

