% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsets.R
\name{CTMC.line.sample}
\alias{CTMC.line.sample}
\title{Sample the number of gaining traits on a line based on CTMC}
\usage{
CTMC.line.sample(x0, t, k12, k21)
}
\arguments{
\item{x0}{Start point from CTMC}

\item{t}{Total length of time}

\item{k12}{transition rate from state 0 to state 1}

\item{k21}{transition rate from state 1 to state 0}
}
\value{
number of gain traits and the ending state
}
\description{
Sample the number of gaining traits on a line based on CTMC
}

