% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMI_thres_posterior.R
\name{fastpost2}
\alias{fastpost2}
\title{Posterior probability for the number of crossing given the laten value on each node}
\usage{
fastpost2(data, phy, epsilon, k = 0)
}
\arguments{
\item{phy}{A phylo type object.}

\item{epsilon}{the radius of the interval for crossing}

\item{d}{the discrete observation on the tip of the tree.}

\item{root.prior}{A numeric vector with two values, which is the mean and variance for the prior distribution (normal) for the root liability.}
}
\value{
A vector associate with the posterior probabililty for the number of upcrossing
}
\description{
Posterior probability for the number of crossing given the laten value on each node
}

