% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brown_tree_prior1.R
\name{fills_in}
\alias{fills_in}
\title{Generating the number of crossings given the nodes liabilties of the phylogenetic tree}
\usage{
fills_in(phy, data, epsilon, thed = 0, cut = 0, less = F)
}
\arguments{
\item{phy}{A phylo type object.}

\item{data}{The liabilities of the phylogenetic tree nodes.}

\item{epsilon}{the radius of the interval for crossing.}

\item{thed}{the threshold value for Brownian motion threshold model.}

\item{cut}{cutoff value: During the accumulating process, if the number of crossing is greater than the cutoff value, it will stop this iteration and yeild greater than cutoff.}
}
\value{
A vector associate with the prior probabililty for the number of upcrossing
}
\description{
Generating the number of crossings given the nodes liabilties of the phylogenetic tree
}

