% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEphy.R
\name{MLEphy.log}
\alias{MLEphy.log}
\title{Calculate the two-state CTMC likelihood with rate in log parametrization}
\usage{
MLEphy.log(par, data)
}
\arguments{
\item{par}{Numeric vector for the log transition rates}

\item{data}{List phy for phylogenetic tree, tip01 by binary states encoded in 0,1}
}
\value{
CTMC likelihood
}
\description{
Calculate the two-state CTMC likelihood with rate in log parametrization
}

