% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossing.R
\name{OUbridge}
\alias{OUbridge}
\title{simulating and counting the sample path of OU bridge or Brownian bridge}
\usage{
OUbridge(x0, xT, TT, n, theta, mu, sigma, pl = F)
}
\arguments{
\item{x0}{the starting point at t=0}

\item{TT}{the total time}

\item{n}{the number of the grid}

\item{theta}{Default as 0, which means brownian motion}

\item{sigma}{The diffusion parameter of OU process}

\item{pl}{Whether plot the path, default as false}

\item{XT}{the ending point at t=TT}
}
\value{
a vector of OU-bridge
}
\description{
simulating and counting the sample path of OU bridge or Brownian bridge
}

