% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossing.R
\name{OUpath}
\alias{OUpath}
\title{simulate the sample path of OU-process or Brownian motion}
\usage{
OUpath(x0, TT, n, theta = 0, mu, sigma = 1, pl = F)
}
\arguments{
\item{x0}{the starting point at t=0}

\item{TT}{the total time}

\item{n}{the number of the grid}

\item{theta}{Default as 0, which means brownian motion}

\item{sigma}{The diffusion parameter of OU process}

\item{pl}{Whether plot the path, default as false}
}
\value{
A vector of OU-process path or Brownian motion path
}
\description{
simulate the sample path of OU-process or Brownian motion
}

