% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEphy.R
\name{MLEphy}
\alias{MLEphy}
\title{Calculate the two-state CTMC likelihood}
\usage{
MLEphy(par, data)
}
\arguments{
\item{par}{Positive numeric vector for the transition rates}

\item{data}{List phy for phylogenetic tree, tip01 by binary states encoded in 0,1}
}
\value{
CTMC likelihood
}
\description{
Calculate the two-state CTMC likelihood
}

